Pull out X01ScoreRenderer
